Refactor duplicate Emscripten code into a mixin
